LPS-187673 Extract methods from PDFPreviewManagedServiceFactory to PDFPreviewManagedServiceFactoryHelper and apply to usages
