+++ modules/apps/document-library/document-library-preview-pdf/src/main/java/com/liferay/document/library/preview/pdf/internal/PDFProcessorImpl.java
@@ -22,7 +22,7 @@ import com.liferay.document.library.kernel.util.DLPreviewableProcessor;
 import com.liferay.document.library.kernel.util.DLProcessor;
 import com.liferay.document.library.kernel.util.DLUtil;
 import com.liferay.document.library.kernel.util.PDFProcessor;
-import com.liferay.document.library.preview.pdf.internal.configuration.admin.service.PDFPreviewManagedServiceFactory;
+import com.liferay.document.library.preview.pdf.internal.configuration.admin.service.util.PDFPreviewManagedServiceFactoryHelper;
 import com.liferay.document.library.preview.pdf.internal.util.ProcessConfigUtil;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.petra.process.ProcessCallable;
@@ -112,7 +112,7 @@ public class PDFProcessorImpl
 		throws Exception {
 
 		int maxNumberOfPages =
-			_pdfPreviewManagedServiceFactory.getMaxNumberOfPages(
+			_pdfPreviewManagedServiceFactoryHelper.getMaxNumberOfPages(
 				ExtendedObjectClassDefinition.Scope.GROUP.getValue(),
 				destinationFileVersion.getGroupId());
 
@@ -1084,7 +1084,8 @@ public class PDFProcessorImpl
 	private boolean _ghostscriptInitialized;
 
 	@Reference
-	private PDFPreviewManagedServiceFactory _pdfPreviewManagedServiceFactory;
+	private PDFPreviewManagedServiceFactoryHelper
+		_pdfPreviewManagedServiceFactoryHelper;
 
 	@Reference
 	private ProcessExecutor _processExecutor;
