+++ modules/apps/document-library/document-library-preview-pdf/src/main/java/com/liferay/document/library/preview/pdf/internal/configuration/admin/display/BasePDFPreviewConfigurationScreen.java
@@ -15,7 +15,7 @@
 package com.liferay.document.library.preview.pdf.internal.configuration.admin.display;
 
 import com.liferay.configuration.admin.display.ConfigurationScreen;
-import com.liferay.document.library.preview.pdf.internal.configuration.admin.service.PDFPreviewManagedServiceFactory;
+import com.liferay.document.library.preview.pdf.internal.configuration.admin.service.util.PDFPreviewManagedServiceFactoryHelper;
 import com.liferay.document.library.preview.pdf.internal.portlet.action.PDFPreviewConfigurationDisplayContext;
 import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClassDefinition;
 import com.liferay.portal.kernel.language.Language;
@@ -73,7 +73,7 @@ public abstract class BasePDFPreviewConfigurationScreen
 					portal.getLiferayPortletResponse(
 						(PortletResponse)httpServletRequest.getAttribute(
 							JavaConstants.JAVAX_PORTLET_RESPONSE)),
-					pdfPreviewManagedServiceFactory, getScope(),
+					pdfPreviewManagedServiceFactoryHelper, getScope(),
 					_getScopePk(httpServletRequest)));
 
 			RequestDispatcher requestDispatcher =
@@ -93,7 +93,8 @@ public abstract class BasePDFPreviewConfigurationScreen
 	protected Language language;
 
 	@Reference
-	protected PDFPreviewManagedServiceFactory pdfPreviewManagedServiceFactory;
+	protected PDFPreviewManagedServiceFactoryHelper
+		pdfPreviewManagedServiceFactoryHelper;
 
 	@Reference
 	protected Portal portal;
